re format code and java type length 不要转译层，使用多种类适配方式,去除字节编码方式属性

+++ src/main/java/me/snake/tools/utils/BCDByteUtil.java
@@ -12,7 +12,7 @@ public class BCDByteUtil {
         return (byte) ((high << 4) | low);
     }
 
-    public static byte[] number2bcd(long number, byte length) {
+    public static byte[] number2bcd(long number, int length) {
         assert length > 0 && length <= 20;
         byte[] bytes = new byte[length];
         long factor = number;
